backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
alembic
asyncpg
pydantic>=2.0.0
pydantic-settings
python-dotenv